chore(pipnn): u32 Edge, fix doc lies, tracing discipline
Edge type unified to u32:
- Edge::src/dst: usize → u32 (matches Leaf::indices: Vec<u32> and
  ReservoirEntry::neighbor: u32 that everything around it already uses).
- build_leaf / build_leaf_into / build_leaf_with_buffers take &[u32].
- builder.rs: drops the per-leaf `Vec<usize>` widening alloc that ran
  inside the LEAF_BATCH inner loop on every leaf. At BigANN 10M with
  c_max=256 and ~250k leaves that's 250k transient ~2KB allocs gone.
- hash_prune.rs add_edges_batched: last_src is u32, reservoir lookup
  goes through `as usize` once per edge group instead of edge.dst's
  `as u32` cast on every insert.

Doc lies corrected (these were misleading future debugging):
- builder.rs final_prune_from_candidates: doc claimed iterative-alpha
  + resumable inner loop. Code is single-pass paper Algorithm 2.
- hash_prune.rs HashPruneReservoir: doc claimed sorted Vec with O(log l)
  lookups. Storage is unsorted; lookup is SIMD linear scan via
  find_hash (AVX-512 8-way / AVX2 4-way / scalar).
- hash_prune.rs add_edges_batched: doc claimed it sorts edges. It
  doesn't — it caches the last source lock, which is correct because
  leaf-build emits edges grouped by src. Documented the invariant.
- builder.rs PiPNNGraph::search comment "Greedy beam search" was
  actually best-first with sorted candidate list — fixed.

Tracing / log discipline:
- builder.rs: 4 `eprintln!("SIMD: ...")` collapsed into a single
  `log_simd_tier()` helper that emits one tracing::info! line.
- builder.rs final_prune input/output: 2 `println!` blocks converted
  to tracing::info! so the output doesn't pollute stdout when PiPNN
  runs inside the disk builder.
- partition.rs `assert!(iteration <= 50)` panic message now includes
  npoints / c_max / c_min / fanout for debuggability when (if ever)
  the partition diverges.

Distance dispatcher fix:
- make_dist_fn now takes ndims and threads it to
  `DistanceProvider::distance_comparer(metric, Some(ndims))` — the
  dimension-specialized SIMD kernel was previously unavailable to
  find_medoid and the test-only PiPNNGraph::search because they
  passed None.

Non-asserting tests promoted to actual assertions:
- test_lsh_sketches_different_seeds: now checks at least one pair
  hash differs (statistical, but deterministic across many pairs).
- test_relative_hash_symmetry_broken (renamed
  test_relative_hash_is_asymmetric): now `assert_ne!`.

Validated:
- cargo clippy --all-targets -- -D warnings: clean.
- cargo test --lib: 91/91 pass.

+++ diskann-pipnn/benches/pipnn_bench.rs
@@ -80,7 +80,7 @@ fn bench_build_leaf(c: &mut Criterion) {
 
     for &(n, ndims, k) in &[(128, 128, 3), (512, 128, 4), (1024, 128, 4), (512, 384, 5)] {
         let data = random_data(n, ndims, 42);
-        let indices: Vec<usize> = (0..n).collect();
+        let indices: Vec<u32> = (0..n as u32).collect();
 
         group.throughput(Throughput::Elements(n as u64));
         group.bench_with_input(
@@ -112,7 +112,7 @@ fn bench_hash_prune_add_edges(c: &mut Criterion) {
         let leaf_size = 512;
         let k = 4;
         let leaf_data = random_data(leaf_size, ndims, 99);
-        let leaf_indices: Vec<usize> = (0..leaf_size).collect();
+        let leaf_indices: Vec<u32> = (0..leaf_size as u32).collect();
         let edges = leaf_build::build_leaf(&leaf_data, ndims, &leaf_indices, k, Metric::L2);
 
         group.throughput(Throughput::Elements(edges.len() as u64));
